requirements.txt
Flask
Flask-Bootstrap
transformers
Pillow